requirements.txt
streamlit
python-nmap
openai
requests 